feat(wasip2,process,io): add JSPI/asyncify async mode support
Make WASI P2 shim, process manager, and I/O layer support both
synchronous (Workers + SharedArrayBuffer) and asynchronous
(asyncify/JSPI) execution modes via a generic MaybePromise<T, Sync>
type parameter that propagates through all interfaces and classes.

Key changes:
- MaybePromise<T, Sync extends boolean> conditional type with
  isThenable, chainMaybePromise, mapMaybePromise helpers
- InputStream<Sync>, OutputStream<Sync>, Pollable<Sync>,
  Descriptor<Sync>, TcpSocket<Sync>, UdpSocket<Sync> generic classes
- InputStreamHandler<Sync>, OutputStreamHandler<Sync>,
  DescriptorHandler<Sync>, FileSystemProvider<Sync>,
  SocketProvider<Sync>, HttpClient<Sync> generic interfaces
- Unified FsDescriptorHandler<Sync> replacing separate sync/async
  handlers with shared chainMaybePromise/mapMaybePromise helpers
- poll() uses Promise.race() for async pollables
- WASIShim async mode with setTimeout-based clock subscriptions
- SimpleProcessManager async wait via Promise-returning handlers
- Async queue pipe with Promise-based blockingRead
- TCP/UDP stream I/O wired to provider receive/send
- IP literal address parsing (IPv4/IPv6 with :: expansion)
- HTTP trailer support and timeout propagation
- Triple-build transpile scripts (sync/jspi/asyncify) for shell,
  coreutils, and example packages
- NodeSocketProvider: shutdown types, socket options, IPv6 DNS
- Comprehensive test coverage (1883 tests, 0 failures)

+++ packages/process/src/io/pipes.test.ts
@@ -209,7 +209,7 @@ describe('SharedPipe stress', () => {
     output[Symbol.dispose]();
     try {
       while (true) {
-        const result = input.blockingRead(256n);
+        const result = input.blockingRead(256n) as Uint8Array;
         for (const b of result) read.push(b);
       }
     } catch { /* closed */ }
